requirements.txt
inference-cli
numpy
Pillow
requests
